requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
av==15.0.0
blessed==1.21.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
ctranslate2==4.6.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
editor==1.6.6
executing==2.2.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.1.8
html2image==2.0.7
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
inquirer==3.4.1
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
litellm==1.76.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.3.2
onnxruntime==1.22.1
open-interpreter==0.4.3
openai==1.101.0
outcome==1.3.0.post0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
piper-tts==1.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.2
readchar==4.2.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rich==13.9.4
rpds-py==0.27.0
rsa==4.9.1
runs==1.2.2
scipy==1.16.1
selenium==4.35.0
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
stack-data==0.6.3
starlette==0.37.2
sympy==1.14.0
termcolor==2.3.0
tiktoken==0.7.0
tokenizers==0.21.4
tokentrim==0.1.13
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
wget==3.2
wsproto==1.2.0
xmod==1.8.1
yarl==1.20.1
yaspin==3.1.0
zipp==3.23.0
